requirements.txt
django
djangorestframework
coreapi
setuptools
mysqlclient
djangorestframework-simplejwt